Twitter.find_or_create supports merging accounts.
What happens if a user signs in via twitter, when he is already 
logged in as a user, and the twitter handle exists already in 
the database, but belongs to another user? In this case we take
away the other user's twitter identity and put it to this user.

This scenario might sound esoteric. A use case, however, is:

- user logs in via its twitter @handle
- user logs out
- now user registered via "user@email.com" email
- user starts or shares a quest. The application asks the user 
  to add a twitter login.
- user enters his @twitter handle.