Fixes twitter auth
To be honest, I have no idea if the change in the middleware code or the updated Gemfile fixes twitter login. I also have no idea why the old code
still works with an old twitter application record (the one at staging) but not on live, which was just created today.

Anyways: the updated code works on my machine in a live instance, and hopefully also when deployed.